requirements_enterprise.txt
# Enterprise SCADA AI System Requirements

# Core FastAPI and Web Framework
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
jinja2==3.1.2

# Database Connectivity
mysql-connector-python==8.2.0
pymongo==4.6.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Data Science and AI
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2
scipy==1.11.4

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Authentication and Security
pyjwt[crypto]==2.8.0
bcrypt==4.1.2
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Caching and Session Management
redis==5.0.1
hiredis==2.2.3

# Monitoring and Logging
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0
configparser==6.0.0

# Communication and Notifications
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
smtplib2==0.2.1
slack-sdk==3.26.1

# Deployment and Containerization
docker==6.1.3
docker-compose==1.29.2

# Task Scheduling
schedule==1.2.0
celery[redis]==5.3.4

# File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.22.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Documentation
swagger-ui-bundle==0.0.9
redoc==2.1.0

# Performance and Optimization
gunicorn==21.2.0
gevent==23.9.1